requirements.txt
paho-mqtt==1.4.0
redis==3.2.1
walrus==0.7.1
